// packages/api/migrations/snapshots/fdfee8a763c0359398d8f6ba36c4b27d120781361425e275b3f2bc693546ea55/contract.d.ts
export type CodecTypes = PgTypes;
export type LaneCodecTypes = CodecTypes;
export type QueryOperationTypes = PgAdapterQueryOps<CodecTypes>;
// ...
export type FieldOutputTypes = {
  readonly public: {
    readonly Comment: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly comment: CodecTypes['pg/text@1']['output'];
      readonly created_at: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly userId: CodecTypes['pg/text@1']['output'];
      readonly postId: CodecTypes['pg/text@1']['output'];
    };
    readonly Follow: {
      readonly followerId: CodecTypes['pg/text@1']['output'];
      readonly followingId: CodecTypes['pg/text@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
    };
    readonly Like: {
      readonly userId: CodecTypes['pg/text@1']['output'];
      readonly postId: CodecTypes['pg/text@1']['output'];
    };
    readonly Post: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly title: CodecTypes['pg/text@1']['output'];
      readonly content: CodecTypes['pg/text@1']['output'];
      readonly slug: CodecTypes['pg/text@1']['output'];
      readonly created_at: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly published_at: CodecTypes['pg/timestamptz-temporal@1']['output'] | null;
      readonly authorId: CodecTypes['pg/text@1']['output'];
    };
    readonly Post_Tag: {
      readonly postId: CodecTypes['pg/text@1']['output'];
      readonly tagId: CodecTypes['pg/text@1']['output'];
    };
    readonly Tag: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
    };
    readonly User: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
      readonly email: CodecTypes['pg/text@1']['output'];
      readonly bio: CodecTypes['pg/text@1']['output'] | null;
      readonly password: CodecTypes['pg/text@1']['output'];
    };
  };
};
export type FieldInputTypes = {
  readonly public: {
    readonly Comment: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly comment: CodecTypes['pg/text@1']['input'];
      readonly created_at: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly userId: CodecTypes['pg/text@1']['input'];
      readonly postId: CodecTypes['pg/text@1']['input'];
    };
    readonly Follow: {
      readonly followerId: CodecTypes['pg/text@1']['input'];
      readonly followingId: CodecTypes['pg/text@1']['input'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
    };
    readonly Like: {
      readonly userId: CodecTypes['pg/text@1']['input'];
      readonly postId: CodecTypes['pg/text@1']['input'];
    };
    readonly Post: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly title: CodecTypes['pg/text@1']['input'];
      readonly content: CodecTypes['pg/text@1']['input'];
      readonly slug: CodecTypes['pg/text@1']['input'];
      readonly created_at: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly published_at: CodecTypes['pg/timestamptz-temporal@1']['input'] | null;
      readonly authorId: CodecTypes['pg/text@1']['input'];
    };
    readonly Post_Tag: {
      readonly postId: CodecTypes['pg/text@1']['input'];
      readonly tagId: CodecTypes['pg/text@1']['input'];
    };
    readonly Tag: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
    };
    readonly User: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
      readonly email: CodecTypes['pg/text@1']['input'];
      readonly bio: CodecTypes['pg/text@1']['input'] | null;
      readonly password: CodecTypes['pg/text@1']['input'];
    };
  };
};
export type StorageColumnTypes = {
  readonly public: {
    readonly comment: {
      readonly comment: CodecTypes['pg/text@1']['output'];
      readonly created_at: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly postId: CodecTypes['pg/text@1']['output'];
      readonly userId: CodecTypes['pg/text@1']['output'];
    };
    readonly follow: {
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly followerId: CodecTypes['pg/text@1']['output'];
      readonly followingId: CodecTypes['pg/text@1']['output'];
    };
    readonly like: {
      readonly postId: CodecTypes['pg/text@1']['output'];
      readonly userId: CodecTypes['pg/text@1']['output'];
    };
    readonly post: {
      readonly authorId: CodecTypes['pg/text@1']['output'];
      readonly content: CodecTypes['pg/text@1']['output'];
      readonly created_at: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly published_at: CodecTypes['pg/timestamptz-temporal@1']['output'] | null;
      readonly slug: CodecTypes['pg/text@1']['output'];
      readonly title: CodecTypes['pg/text@1']['output'];
    };
    readonly post_Tag: {
      readonly postId: CodecTypes['pg/text@1']['output'];
      readonly tagId: CodecTypes['pg/text@1']['output'];
    };
    readonly tag: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
    };
    readonly user: {
      readonly bio: CodecTypes['pg/text@1']['output'] | null;
      readonly email: CodecTypes['pg/text@1']['output'];
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
      readonly password: CodecTypes['pg/text@1']['output'];
    };
  };
};
export type StorageColumnInputTypes = {
  readonly public: {
    readonly comment: {
      readonly comment: CodecTypes['pg/text@1']['input'];
      readonly created_at: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly postId: CodecTypes['pg/text@1']['input'];
      readonly userId: CodecTypes['pg/text@1']['input'];
    };
    readonly follow: {
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly followerId: CodecTypes['pg/text@1']['input'];
      readonly followingId: CodecTypes['pg/text@1']['input'];
    };
    readonly like: {
      readonly postId: CodecTypes['pg/text@1']['input'];
      readonly userId: CodecTypes['pg/text@1']['input'];
    };
    readonly post: {
      readonly authorId: CodecTypes['pg/text@1']['input'];
      readonly content: CodecTypes['pg/text@1']['input'];
      readonly created_at: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly published_at: CodecTypes['pg/timestamptz-temporal@1']['input'] | null;
      readonly slug: CodecTypes['pg/text@1']['input'];
      readonly title: CodecTypes['pg/text@1']['input'];
    };
    readonly post_Tag: {
      readonly postId: CodecTypes['pg/text@1']['input'];
      readonly tagId: CodecTypes['pg/text@1']['input'];
    };
    readonly tag: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
    };
    readonly user: {
      readonly bio: CodecTypes['pg/text@1']['input'] | null;
      readonly email: CodecTypes['pg/text@1']['input'];
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
      readonly password: CodecTypes['pg/text@1']['input'];
    };
  };
};
// ...
export type TypeMaps = TypeMapsType<
  CodecTypes,
  QueryOperationTypes,
  FieldOutputTypes,
  FieldInputTypes,
  StorageColumnTypes,
  StorageColumnInputTypes,
  AggregateTypes
>;
// ...
export type Contract = ContractWithTypeMaps<ContractBase, TypeMaps>;